sentinelai/requirements.txt
ultralytics>=8.3.0
openai>=1.40.0
streamlit>=1.35.0
opencv-python>=4.9.0
numpy>=1.26.0
pydantic>=2.7.0
python-dotenv>=1.0.0
Pillow>=10.3.0
